moving things around to make room for more instructions
