+++ Cargo.toml
@@ -18,6 +18,9 @@ tape-client = { version = "0.2.0", path = "./client" }
 tape-network = { version = "0.2.0", path = "./network" }
 tape-program = { version = "0.2.0", path = "./program" }
 
+crankx = { version = "0.2.2", features = ["solana"] }
+packx = { version = "0.1.3", features = ["solana"] }
+
 # program dependencies
 bytemuck = "1.14.3"
 bytemuck_derive = "1.7.0"
@@ -25,7 +28,6 @@ num_enum = "0.7.2"
 brine-tree = { version = "0.5.4", features = ["solana"] }
 array-const-fn-init = "0.1.1"
 const-crypto = "0.3.0"
-crankx = { version = "0.2.2", features = ["solana"] }
 
 # solana-specific
 solana-client = "=2.1"
