+++ api/src/instruction/mod.rs
@@ -1,9 +1,4 @@
-mod program;
-mod bin;
-mod miner;
-mod tape;
-
-pub use program::*;
-pub use bin::*;
-pub use miner::*;
-pub use tape::*;
+pub mod program;
+pub mod bin;
+pub mod miner;
+pub mod tape;
