+++ api/src/instruction/tape.rs
@@ -1,5 +1,9 @@
 use steel::*;
-use crate::consts::*;
+use crate::{
+    consts::*,
+    pda::*,
+    utils,
+};
 
 #[repr(u8)]
 #[derive(Clone, Copy, Debug, Eq, PartialEq, TryFromPrimitive)]
@@ -54,5 +58,139 @@ pub struct SetHeader {
 #[derive(Clone, Copy, Debug, Pod, Zeroable)]
 pub struct Subsidize {
     pub amount: [u8; 8],
+}
+
+
+pub fn build_create_ix(
+    signer: Pubkey,
+    name: &str,
+) -> Instruction {
+    let name = utils::to_name(name);
+
+    let (tape_address, _tape_bump) = tape_pda(signer, &name);
+    let (writer_address, _writer_bump) = writer_pda(tape_address);
+
+    Instruction {
+        program_id: crate::ID,
+        accounts: vec![
+            AccountMeta::new(signer, true),
+            AccountMeta::new(tape_address, false),
+            AccountMeta::new(writer_address, false),
+            AccountMeta::new_readonly(solana_program::system_program::ID, false),
+            AccountMeta::new_readonly(sysvar::rent::ID, false),
+            AccountMeta::new_readonly(sysvar::slot_hashes::ID, false),
+        ],
+        data: Create {
+            name,
+        }.to_bytes(),
+    }
+}
+
+pub fn build_set_header_ix(
+    signer: Pubkey,
+    tape: Pubkey,
+    header: &[u8; HEADER_SIZE],
+) -> Instruction {
+    Instruction {
+        program_id: crate::ID,
+        accounts: vec![
+            AccountMeta::new(signer, true),
+            AccountMeta::new(tape, false),
+        ],
+        data: SetHeader {
+            header: *header,
+        }.to_bytes(),
+    }
+}
+
+pub fn build_write_ix(
+    signer: Pubkey,
+    tape: Pubkey,
+    writer: Pubkey,
+    data: &[u8],
+) -> Instruction {
+
+    let mut ix_data = Write{}.to_bytes();
+    ix_data.extend_from_slice(data);
+
+    Instruction {
+        program_id: crate::ID,
+        accounts: vec![
+            AccountMeta::new(signer, true),
+            AccountMeta::new(tape, false),
+            AccountMeta::new(writer, false),
+        ],
+        data: ix_data,
+    }
+}
+
+pub fn build_update_ix(
+    signer: Pubkey,
+    tape: Pubkey,
+    writer: Pubkey,
+    segment_number: u64,
+    old_data: [u8; SEGMENT_SIZE],
+    new_data: [u8; SEGMENT_SIZE],
+    proof: [[u8;32]; SEGMENT_PROOF_LEN],
+) -> Instruction {
+
+    let segment_number = segment_number.to_le_bytes();
+
+    Instruction {
+        program_id: crate::ID,
+        accounts: vec![
+            AccountMeta::new(signer, true),
+            AccountMeta::new(tape, false),
+            AccountMeta::new(writer, false),
+        ],
+        data: Update {
+            segment_number,
+            old_data,
+            new_data,
+            proof,
+        }.to_bytes(),
+    }
+}
+
+pub fn build_finalize_ix(
+    signer: Pubkey, 
+    tape: Pubkey,
+    writer: Pubkey,
+) -> Instruction {
+
+    Instruction {
+        program_id: crate::ID,
+        accounts: vec![
+            AccountMeta::new(signer, true),
+            AccountMeta::new(tape, false),
+            AccountMeta::new(writer, false),
+            AccountMeta::new(ARCHIVE_ADDRESS, false),
+            AccountMeta::new_readonly(solana_program::system_program::ID, false),
+            AccountMeta::new_readonly(sysvar::rent::ID, false),
+        ],
+        data: Finalize {}.to_bytes(),
+    }
+}
+
+pub fn build_subsidize_ix(
+    signer: Pubkey, 
+    ata: Pubkey,
+    tape: Pubkey,
+    amount: u64,
+) -> Instruction {
+
+    Instruction {
+        program_id: crate::ID,
+        accounts: vec![
+            AccountMeta::new(signer, true),
+            AccountMeta::new(ata, false),
+            AccountMeta::new(tape, false),
+            AccountMeta::new(TREASURY_ATA, false),
+            AccountMeta::new_readonly(spl_token::ID, false),
+        ],
+        data: Subsidize {
+            amount: amount.to_le_bytes(),
+        }.to_bytes(),
+    }
 }
