+++ api/src/lib.rs
@@ -5,7 +5,6 @@ pub mod instruction;
 pub mod loaders;
 pub mod pda;
 pub mod rent;
-pub mod sdk;
 pub mod state;
 pub mod types;
 pub mod utils;
@@ -17,11 +16,9 @@ pub mod prelude {
     pub use crate::consts::*;
     pub use crate::error::*;
     pub use crate::event::*;
-    pub use crate::instruction::*;
     pub use crate::loaders::*;
     pub use crate::pda::*;
     pub use crate::rent::*;
-    pub use crate::sdk::*;
     pub use crate::state::*;
     pub use crate::types::*;
     pub use crate::utils::*;
