+++ api/src/state/bin.rs
@@ -8,8 +8,9 @@ use super::AccountType;
 pub struct Bin {
     pub number: u64,
 
-    pub miner: Pubkey,
-    pub state: TapeTree, 
+    pub miner:    Pubkey,
+    pub state:    TapeTree, 
+    pub contains: [u8; 32], 
 
     pub last_proof_block: u64,
     pub last_proof_at: i64,
