+++ client/src/tape/subsidize.rs
@@ -7,7 +7,7 @@ use solana_sdk::{
     transaction::Transaction,
     pubkey::Pubkey,
 };
-use tape_api::prelude::*;
+use tape_api::instruction::tape::build_subsidize_ix;
 use solana_client::nonblocking::rpc_client::RpcClient;
 use crate::utils::*;
