+++ program/src/tape/create.rs
@@ -1,4 +1,5 @@
 use tape_api::prelude::*;
+use tape_api::instruction::tape::Create;
 use solana_program::{
     keccak::hashv,
     slot_hashes::SlotHash,
